Add files via upload
This program in asp.net to create a webform application with a form and 2 listboxes(1stcolorsUppercase and 1stcolorsLowercase ) placed horizontally.
add a class rainbow that has an array of string representing the 7 colors in the rainbow.
Implement the IEnumerable interface on the rainbow class and yield return the colors using a for loop.
In the load event of the form, create an instance of rainbow class. create a reference of type IEnumerator and 
assign it an enumerator by calling the rainbow objects GetEnumerator() method.Add the current item to 1stColorsUppercase 
listbox by converting the item to an uppercase string within a while loop that uses the enumerator.